// FrontGestionFacturas/src/app/models/interfaces/user-token.interface.ts
// Generated by https://quicktype.io

import { UserRol } from "./user.interface";

export interface IToken {
    token:    IUserToken;
    userName: string;
}

export interface IUserToken {
    id:           number;
    token:        string;
    userName:     string;
    userRol:      UserRol;
    validity:     string;
    refreshToken: string;
    emailId:      string;
    guidId:       string;
    expireTime:   Date;
}